# Conflicts:
#	.github/workflows/unix.yml
#	base/core/include/motis/core/schedule/station.h

+++ base/core/include/motis/core/access/realtime_access.h
@@ -26,6 +26,6 @@ ev_key const& get_current_ev_key(schedule const&, ev_key const&);
 
 ev_key const& get_orig_ev_key(schedule const&, ev_key const&);
 
-int get_schedule_track(schedule const&, ev_key const&);
+uint16_t get_schedule_track(schedule const&, ev_key const&);
 
 }  // namespace motis
